requirements.txt


Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
chromadb>=0.4.0
PyPDF2>=3.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.0
